requirements.txt
behave==1.2.6
requests==2.31.0
python-dotenv==1.0.0

mysql-connector-python~=8.1.0